+++ server/controllers/utils/register-hardware.js
